# backend_go/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkyViewLab Go Control Plane API",
    "version": "1.0.0-gate-b",
    "description": "Machine-checkable contract for the current Go /api/v1 control plane. This contract describes the implemented Gate B boundary; it does not claim that every workbench domain schema or production NO-GO item is complete. All JSON success responses use {data, requestId}; all JSON failures use {error:{code,message}, message, requestId}."
  },
  "servers": [
    {"url": "https://api.example.com", "description": "Production placeholder; deployment supplies the real origin"},
    {"url": "http://127.0.0.1:8080", "description": "Local development"}
  ],
...
  "paths": {
...
    "/api/v1/tools/catalog": {
      "get": {"operationId": "listToolCapabilities", "tags": ["tools"], "summary": "Public server-authoritative capability catalog", "description": "Go intersects Python capabilities with its own conservative execution gates. Missing or invalid upstream data fails closed.", "security": [], "x-max-upstream-response-bytes": 4194304, "responses": {"200": {"$ref": "#/components/responses/CapabilityList"}, "502": {"$ref": "#/components/responses/Error502"}}}
    },
    "/api/v1/tools/{slug}/run": {
      "parameters": [{"$ref": "#/components/parameters/Slug"}],
      "post": {"operationId": "rejectDirectToolRun", "tags": ["tools"], "summary": "Reject direct tool execution", "description": "Frozen rejection endpoint. Clients must create a project job instead.", "security": [{"sessionCookie": []}], "parameters": [{"$ref": "#/components/parameters/Origin"}], "x-origin-required-for-cookie": true, "x-execution-policy": "always-rejected", "responses": {"401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "409": {"description": "Always rejected with error.code JOB_API_REQUIRED", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}, "examples": {"rejected": {"value": {"error": {"code": "JOB_API_REQUIRED", "message": "计算必须通过项目作业接口创建，客户端不能直接执行并上报结果"}, "message": "计算必须通过项目作业接口创建，客户端不能直接执行并上报结果", "requestId": "request-id"}}}}}}, "500": {"$ref": "#/components/responses/Error500"}}}
    },
    "/api/v1/analysis/text": {
      "post": {"operationId": "rejectLegacyTextAnalysis", "tags": ["tools"], "summary": "Reject legacy direct analysis", "description": "Frozen rejection endpoint retained for compatibility; clients must use project jobs.", "security": [{"sessionCookie": []}], "parameters": [{"$ref": "#/components/parameters/Origin"}], "x-origin-required-for-cookie": true, "x-execution-policy": "always-rejected", "responses": {"401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "409": {"description": "Always rejected with error.code JOB_API_REQUIRED", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}, "500": {"$ref": "#/components/responses/Error500"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "skyviewlab_session", "description": "Opaque HttpOnly BFF session. Browser unsafe operations also require an allowlisted Origin."},
      "oidcStateCookie": {"type": "apiKey", "in": "cookie", "name": "skyviewlab_oidc_state", "description": "Short-lived, one-time OIDC correlation cookie; not an authenticated application session."},
      "workerBearer": {"type": "http", "scheme": "bearer", "bearerFormat": "opaque", "description": "A distinct opaque token bound server-side to one worker ID and 1..64 exact tenantId/workspaceId/single-slug/action-allowlist scopes. Empty scopes, wildcards and duplicate boundaries are rejected. Production loads the bounded credential set from WORKER_CREDENTIALS_FILE; short-lived workload JWT or mTLS remains a deployment hardening boundary."}
    },
    "parameters": {
      "Origin": {"name": "Origin", "in": "header", "required": true, "description": "Exactly one browser Origin, matching the configured allowlist byte-for-byte.", "schema": {"type": "string", "format": "uri", "maxLength": 2048}},
...
      "Slug": {"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128}},
...
    },
    "schemas": {
...
      "RequestID": {"type": "string", "minLength": 1, "maxLength": 128},
      "ErrorDetail": {"type": "object", "additionalProperties": false, "required": ["code", "message"], "properties": {"code": {"type": "string", "minLength": 1}, "message": {"type": "string"}}},
      "ErrorEnvelope": {"type": "object", "additionalProperties": false, "required": ["error", "message", "requestId"], "properties": {"error": {"$ref": "#/components/schemas/ErrorDetail"}, "message": {"type": "string"}, "requestId": {"$ref": "#/components/schemas/RequestID"}}},
      "DataEnvelope": {"type": "object", "required": ["data", "requestId"], "properties": {"data": {}, "requestId": {"$ref": "#/components/schemas/RequestID"}}},
...
      "WorkbenchStatus": {"type": "string", "enum": ["prototype", "planned", "security-blocked"]},
...
      "Capability": {"type": "object", "required": ["slug", "status", "executionAllowed", "productionReady"], "properties": {"slug": {"type": "string"}, "status": {"$ref": "#/components/schemas/WorkbenchStatus"}, "executionAllowed": {"type": "boolean"}, "productionReady": {"const": false}}, "additionalProperties": true},
...
      "CapabilityListEnvelope": {"allOf": [{"$ref": "#/components/schemas/DataEnvelope"}, {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/Capability"}}}}]},
...
    },
    "responses": {
...
      "CapabilityList": {"description": "Capabilities", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CapabilityListEnvelope"}}}},
...
      "Error401": {"description": "Authentication or correlation failed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error403": {"description": "Origin, role, or resource authorization denied", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
      "Error500": {"description": "Internal persistence, encoding, or audit failure", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
      "Error502": {"description": "Identity or Python upstream unavailable/invalid", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
...
    }
  }
}